+++ datasketches/tests/frequencies_test/update.rs
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#![cfg(feature = "frequencies")]
-
 use datasketches::frequencies::ErrorType;
 use datasketches::frequencies::FrequentItemsSketch;
