+++ datasketches/tests/serialization_tests/countmin.rs
@@ -15,17 +15,14 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#![cfg(feature = "countmin")]
-
-#[path = "support/serialization_data.rs"]
-mod serialization_data;
+mod support;
 
 use std::fs;
 
 use datasketches::countmin::CountMinSketch;
 use googletest::assert_that;
 use googletest::prelude::contains_substring;
-use serialization_data::serialization_test_data;
+use support::serialization_test_data;
 
 // This test validates binary format compatibility (deserialize + byte round-trip) for
 // C++ Count-Min snapshots. It intentionally does not assert estimate equivalence against
