+++ datasketches/tests/serialization_tests/theta.rs
@@ -15,10 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#![cfg(feature = "theta")]
-
-#[path = "support/serialization_data.rs"]
-mod serialization_data;
+mod support;
 
 use std::fs;
 use std::path::PathBuf;
@@ -26,7 +23,7 @@ use std::path::PathBuf;
 use datasketches::theta::CompactThetaSketch;
 use googletest::assert_that;
 use googletest::prelude::near;
-use serialization_data::serialization_test_data;
+use support::serialization_test_data;
 
 fn test_sketch_file(path: PathBuf, expected_cardinality: usize, use_compressed_round_trip: bool) {
     let expected = expected_cardinality as f64;
