+++ datasketches/tests/theta_test/sketch.rs
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#![cfg(feature = "theta")]
-
 use datasketches::common::NumStdDev;
 use datasketches::hash_value;
 use datasketches::theta::ThetaSketchBuilder;
